JIN-15 Comment쪽에 권한 관련 로직 추가하기

+++ src/main/java/com/study/jinius/comment/service/CommentService.java
@@ -2,8 +2,10 @@ package com.study.jinius.comment.service;
 
 import com.study.jinius.account.model.Account;
 import com.study.jinius.account.repository.AccountRepository;
+import com.study.jinius.account.service.CustomUserDetailsService;
 import com.study.jinius.comment.model.*;
 import com.study.jinius.comment.repository.CommentRepository;
+import com.study.jinius.common.exception.BadRequestException;
 import com.study.jinius.common.exception.NotFoundException;
 import com.study.jinius.common.model.Status;
 import com.study.jinius.post.model.Post;
@@ -16,16 +18,21 @@ import org.springframework.util.CollectionUtils;
 import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 @RequiredArgsConstructor
 public class CommentService {
+    private final CustomUserDetailsService customUserDetailsService;
     private final AccountRepository accountRepository;
     private final PostRepository postRepository;
     private final CommentRepository commentRepository;
 
+    @Transactional
     public CreateCommentResponse createComment(CommentCreateParam param) {
-        Account account = accountRepository.findById(param.getAccountId())
+        Long accountId = customUserDetailsService.getAccountIdx();
+        Account account = accountRepository.findById(accountId)
+                .filter(Account::isValidAccount)
                 .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
 
         Comment comment = param.toComment(account);
@@ -44,10 +51,15 @@ public class CommentService {
         return CreateCommentResponse.from(result);
     }
 
-
     @Transactional
     public CommentUpdateResponse updateComment(Long idx, PostUpdateParam param) {
-        Comment comment = commentRepository.findById(idx).orElseThrow();
+        Long accountId = customUserDetailsService.getAccountIdx();
+        checkIsValidAccountId(accountId);
+
+        Comment comment = commentRepository.findById(idx)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 댓글입니다."));
+
+        checkCommentOwner(accountId, comment.getAccount().getIdx());
         comment.updateComment(param.getContent(), param.getStatus());
 
         Post post = comment.getPost();
@@ -59,23 +71,40 @@ public class CommentService {
         return CommentUpdateResponse.from(result);
     }
 
-
+    @Transactional
     public Long deleteComment(Long idx) {
-        Comment comment = commentRepository.findById(idx).orElseThrow();
-        comment.deleteComment();
-        Comment result = commentRepository.save(comment);
+        Long accountId = customUserDetailsService.getAccountIdx();
+        checkIsValidAccountId(accountId);
 
+        Comment comment = commentRepository.findById(idx)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 댓글입니다."));
+
+        Post post = comment.getPost();
+        if (Status.DELETED.equals(post.getStatus())) {
+            throw new NotFoundException("삭제된 게시물입니다.");
+        }
+
+        checkCommentOwner(accountId, comment.getAccount().getIdx());
+        comment.deleteComment();
+
+        Comment result = commentRepository.save(comment);
         return result.getIdx();
     }
 
     public List<CommentResponse> getList(Long postIdx) {
-        Post post = postRepository.findById(postIdx).orElseThrow();
+        Post post = postRepository.findById(postIdx)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 게시물입니다."));
 
         if (Status.DELETED.equals(post.getStatus())) {
             throw new NotFoundException("삭제된 게시물입니다.");
         }
 
-        List<Comment> commentList = commentRepository.findAllByPost(post);
+        Long accountId = customUserDetailsService.getAccountIdx();
+
+        List<Comment> commentList = commentRepository.findAllByPost(post).stream()
+                .filter(c -> Objects.equals(c.getAccount().getIdx(), accountId)
+                        || Status.ACTIVE.equals(c.getStatus()))
+                .toList();
         List<CommentResponse> responseList = new ArrayList<>();
         makeCommentResponseList(commentList, responseList);
 
@@ -92,5 +121,21 @@ public class CommentService {
                 makeCommentResponseList(childList, responseList);
             }
         }
+    }
+
+    private void checkCommentOwner(Long accountIdx, Long commentOwnerId) {
+        if (!Objects.equals(accountIdx, commentOwnerId)) {
+            throw new BadRequestException("게시물 작성자가 아닙니다.");
+        }
+    }
+
+    private void checkIsValidAccountId(Long accountId) {
+        getValidAccount(accountId);
+    }
+
+    private Account getValidAccount(Long accountId) {
+        return accountRepository.findById(accountId)
+                .filter(Account::isValidAccount)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다."));
     }
 }
